Refactor to support relative-time and locale data
**This is a major refactor.**

This has the following changes:

- Add relative format support (e.g., "3 hours ago").
- Add CLDR locale data to this package.
- Set English as the default locale.
- Use `intl-messageformat` with locale data separated from runtime.

Since the locale data has been separated from the runtime, to use this package, a developer must now include two scripts on the page; e.g., for French:

```html
<script src="react-intl/react-intl.min.js"></script>
<script src="react-intl/locale-data/fr.js"></script>
```

The reason for this separation of data from runtime, is file size. These changes bring the whole `react-intl.min.js` bundle down to **6.3k gz**, and with all locales — `react-intl-with-locales.min.js` — it's **33k gz**. Since most apps will only be operating in one locale at a time, it makes sense to extract the data from the library, and include it into the page as a separate `<script>`.

- [ ] Merge https://github.com/yahoo/intl-format-cache/pull/1
- [ ] Upgrade `intl-format-cache` in `package.json`
- [ ] Update README to explain how to load locale data
- [ ] Run the release build (ommitted build for PR)
- [ ] Publish to npm as `1.0.0-rc-2`